# gator-leasing-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for subleasing website ReLease",
        "title": "ReLease API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/leases": {
            "get": {
                "description": "get all leases",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "List leases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewModel.Lease"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Auth0": []
                    }
                ],
                "description": "post a lease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Create a lease",
                "parameters": [
                    {
                        "description": "create lease request",
                        "name": "createLeaseRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewModel.CreateLeaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "id of created lease",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/leases/{id}": {
            "put": {
                "security": [
                    {
                        "Auth0": []
                    }
                ],
                "description": "update a lease by id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Update a lease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "lease id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "edit lease request",
                        "name": "editLeaseRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewModel.EditLeaseRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Auth0": []
                    }
                ],
                "description": "delete a lease by id",
                "tags": [
                    "leases"
                ],
                "summary": "Delete a lease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "lease id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "viewModel.CreateLeaseRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "viewModel.EditLeaseRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "viewModel.Lease": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Auth0": {
            "description": "Auth0 protects our endpoints",
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://dev-nkzmwy1mucvvl5xb.us.auth0.com/oauth/token"
        }
    }
}